fix(lint): workspace clippy -D warnings clean; unbreak petals bench target
Hide Phase 1 finalization pass per HIDE-SPEC.md success criteria:
- drop dangling [[bench]] inference_bench (file never committed) + criterion dep
- fix petals backends: strip_prefix over manual slicing, try_filter_map SSE
  pipeline, clamp, dead vars/imports, unused Instant in chat paths
- client: item-level allow(dead_code) on unwired scaffold modules
  (brain/drive/serverless), remove unused imports, NodeId copy

Verified: cargo test --workspace 114/0, clippy --all-targets -D warnings
exit 0, scripts/hide-smoke.sh E2E green (tunnel + relay blind audit).

+++ crates/petals/src/lib.rs
@@ -229,7 +229,7 @@ impl PetalsEngine {
         total_ram_gb *= 1.0 + (concurrent * 0.1); // 10% per concurrent request
 
         Ok(ResourceRequest {
-            cpu_cores: max_cpu_cores.max(1) as u32,
+            cpu_cores: max_cpu_cores.max(1),
             ram_gb: total_ram_gb.max(1.0),
             gpu_vram_gb: if gpu_vram_gb > 0.0 { Some(gpu_vram_gb) } else { None },
             concurrent_requests: self.config.max_concurrent_requests as u32,
